requirements.txt
Pillow
django-rest-framework
psycopg2
dj_database_url
gunicorn
django-heroku
django-cors-headers
cloudinary
djangorestframework-simplejwt
